+++ js/form.js
@@ -16,9 +16,7 @@ const conditions = {
 };
 
 const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
-const errorMessageElement = errorMessageTemplate.cloneNode(true);
 const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
-const successMessageElement = successMessageTemplate.cloneNode(true);
 const adForm = document.querySelector('.ad-form');
 const fieldsets = adForm.querySelectorAll('fieldset');
 const mapFilters = document.querySelector('.map__filters');
@@ -31,34 +29,27 @@ const roomCapacitySelect = adForm.querySelector('#capacity');
 const timeInSelect = adForm.querySelector('#timein');
 const timeOutSelect = adForm.querySelector('#timeout');
 
-const setInactive = () => {
-  adForm.classList.add('ad-form--disabled');
+const toggleState = (flag) => {
+  adForm.classList.toggle('ad-form--disabled');
   for (const fieldset of fieldsets) {
-    fieldset.disabled = true;
+    fieldset.disabled = flag;
   }
-  mapFilters.classList.add('map__filters--disabled');
+  mapFilters.classList.toggle('map__filters--disabled');
   for (const select of selects) {
-    select.disabled = true;
+    select.disabled = flag;
   }
+};
+
+const setInactive = () => {
+  toggleState(true);
 };
 
 const setActive = () => {
-  adForm.classList.remove('ad-form--disabled');
-  for (const fieldset of fieldsets) {
-    fieldset.disabled = false;
-  }
-  mapFilters.classList.remove('map__filters--disabled');
-  for (const select of selects) {
-    select.disabled = false;
-  }
+  toggleState(false);
 };
 
 const removeMessage = () => {
-  if (document.body.contains(successMessageElement)) {
-    successMessageElement.remove();
-  } else if (document.body.contains(errorMessageElement)) {
-    errorMessageElement.remove();
-  }
+  document.querySelectorAll('.success, .error').forEach((messageElement) => messageElement.remove());
 };
 
 titleInput.addEventListener('input', () => {
@@ -116,5 +107,4 @@ document.addEventListener('click', () => {
   removeMessage();
 });
 
-
-export {setInactive, setActive, adForm, successMessageElement, errorMessageElement};
+export {setInactive, setActive, adForm, successMessageTemplate, errorMessageTemplate};
